fixed developing state
- added possibility to create and delete for files and folders on the desktop
- added open possibility for dirs on the desktop
- remove context menu from clicking in the modal window
- added new properties to the context reducer
- added difference by context menu for items and free space folders

+++ client/src/app/mod/mod.component.ts
@@ -393,8 +393,11 @@ export class ModComponent implements OnInit, AfterViewInit {
 		//find element with class context-menu
 		let contextMenu = document.getElementsByClassName('context-menu');
 		//disable this element
-		contextMenu[0].remove();
-		this.contextMenuService.removeAllFromContext();
+		if(contextMenu[0] != null){
+			contextMenu[0].remove();
+			this.contextMenuService.removeAllFromContext();
+		}
+
 	}
